chore/deps: update scripts to check, build and test auth_ffi
