+++ scripts/build-mock
@@ -6,4 +6,5 @@ cargo build --features=mock-network --release --manifest-path=safe_core/Cargo.to
 cargo build --features="testing mock-network" --release --lib --tests --manifest-path=safe_core/Cargo.toml
 cargo build --features="testing mock-network" --release --lib --tests --manifest-path=safe_authenticator/Cargo.toml
 cargo build --features="testing mock-network" --release --lib --tests --manifest-path=safe_app/Cargo.toml
+cargo build --features="testing mock-network" --release --lib --tests --manifest-path=safe_authenticator_ffi/Cargo.toml
 cargo build --features=mock-network --release --lib --tests --manifest-path=tests/Cargo.toml
