+++ scripts/check-mock
@@ -5,5 +5,6 @@ set -e -x;
 cargo check --features=mock-network --release --manifest-path=safe_core/Cargo.toml
 cargo check --features="testing mock-network" --release --lib --tests --manifest-path=safe_core/Cargo.toml
 cargo check --features="testing mock-network" --release --lib --tests --manifest-path=safe_authenticator/Cargo.toml
+cargo check --features="testing mock-network" --release --lib --tests --manifest-path=safe_authenticator_ffi/Cargo.toml
 cargo check --features="testing mock-network" --release --lib --tests --manifest-path=safe_app/Cargo.toml
 cargo check --features="mock-network" --release --lib --tests --manifest-path=tests/Cargo.toml
