+++ scripts/check-real
@@ -5,5 +5,6 @@ set -e -x;
 cargo check --release --manifest-path=safe_core/Cargo.toml
 cargo check --features=testing --release --lib --tests --manifest-path=safe_core/Cargo.toml
 cargo check --features=testing --release --lib --tests --manifest-path=safe_authenticator/Cargo.toml
+cargo check --features=testing --release --lib --tests --manifest-path=safe_authenticator_ffi/Cargo.toml
 cargo check --features=testing --release --lib --tests --manifest-path=safe_app/Cargo.toml
 cargo check --release --lib --tests --manifest-path=tests/Cargo.toml
