+++ scripts/test-with-mock-vault-file
@@ -5,5 +5,6 @@ set -e -x
 export RUST_BACKTRACE=full
 cargo test --release --features=mock-network --manifest-path=safe_core/Cargo.toml
 cargo test --release --features=mock-network --manifest-path=safe_authenticator/Cargo.toml
+cargo test --release --features=mock-network --manifest-path=safe_authenticator_ffi/Cargo.toml
 cargo test --release --features=mock-network --manifest-path=safe_app/Cargo.toml
 cargo test --release --features=mock-network --manifest-path=tests/Cargo.toml
